Add privilege escalation detection support
This consists of several components:

1) Kernel support for sending notifications on certain events.
2) qemu support for passing those notifications on to a policy agent.
3) A policy agent that builds a tree of process state and verifies that
   this state is unmodified during in-kernel privilege checks.

There's also a minimal amount of code in rkt to add support for this.

Notifications are handled by hitting an io port and passing arguments in
registers. The kernel then blocks the running process until the notification
is handled. This is achieved by the policy agent clearing the blocking flag
and allowing execution to proceed.

Change in privilege triggered by executing a setuid application is handled
specially. --fixed-r12 is passed to the compiler in order to prevent r12
being used in any normal code. The only write instruction to r12 is in an
inline function that is called at the top of execve(), storing the PID.
When the kernel hits the monitor, the monitor ensures that the PID of the
process requesting the privilege change matches the PID in r12, thus
ensuring that the request came via the full exec() call chain. This means
that all appropriate checks in exec() have already passed, making it
difficult for an attacker to trick the monitor into accepting elevated
privileges.

A --monitor flag is added to indicate that this mode should be enabled. If
not passed, qemu will simply instruct the kernel to continue rather than
waiting for monitoring.

One special-cased event is the kernel notifying qemu when it is marking
kernel text read-only. This triggers an mprotect() call in qemu that marks
these pages as read-only in the host page table. Any attempt by the guest
to mark them read-write (in order to disable the syscall notification,
for instance) will trigger a segfault in the host.